Improve definitions and add PHPUnit tests
Distinguish between a key and an index. A key is a single key (i.e.
without any dots or deep indexing), but an index may contain dots
and several nested keys. For example, user.id is an index, not a
key. Because of this, rename almost all methods in JSON class and
also update PHPDoc comments and variables.
Add PHPUnit tests, for testing JSON::get() and JSON::set() methods.