Add JSON::getByReference() method
As user might change data somewhere, it is not guarateed that the
data remains as a recursive array. So, instead of converting every
value to scalar or array, allow everything to be in data.
No errors in JSON::findElementRecursive()? No, change its indexing
method to convert object very effectively, considering the
performance (however, it should be increased in the future).
As a result, remove JSON::setDataTo() and JSON::getOptimalValue()
methods. For keeping the JSON::OPT_JSON_DECODE_ALWAYS option to
function, add JSON::toJsonIfNeeded() to handle it.
Remove JSON::TYPE_FAST also. It was a bad idea at all. Note that
all of these changes is because of the new method of converting to
arrays in JSON::convertToArray(), which is drastically faster than
JSON encoding and decoding.
Also, reorder and rename many things. Some improvements in the
comments. Fix so many bugs. And make some other code improvements.