Change two namespace (BC break)
To be consistent, now everything is under the namespace of
"MAChitgarha\Json".

This is not something really dangerous for breaking the previous
users; as the package is going the rewritten completely, because
the purpose of the package is going to be changed. So, this is non-
sense. Yes, we try to minimize the BC breaks, but, sometimes, a
BC break is better than sticking with current stuff.