Decoded data can be array or object either, from now
Previously, data could only be array. Forcing everything to be an
array leads to significant performance lost, and make things pretty
slow.

However, considering both array and object is allowed, but all
interactors and providers are intended to work with arrays by
default. Nontheless, this should be discussed more in the future.