Remove $data parameter from BaseAdapterInterface's constructor
Previously, the parameter was considered to be held inside the
adapter, and instead of passing data as an argument to needed
methods, that specific data class would be updated. This causes
more ambiguity, as the adapter must always trust that the data is
always updated (possibly by the caller). Also, this makes the
adapters somehow stateful (and may cause scalibility problems).
But, the important thing is, the data should be passed only when
needed. So, the new behaviour should be passing data as input
parameters. It makes the behaviour more stateless, and thus, more
scalable.