Rewrite the entire ProviderSettings class
It now only focuses on holding information about providers,
adapters and their relations with each other and with operations.
It does not create any instances by its own. There would be a
separated class to do that.

Also, split many functionalities into more function, to follow the
rule of simplicity of functions, as the functions have to be as
small as possible. Add a helper class for this purpose, including
ValidationUtil.

Improve documentation, and other small improvements.